requirements.txt
aiogram~=3.1.1
python-dotenv~=1.0.0
gino~=1.0.1
SQLAlchemy~=1.3.24